chore: add basic scaffolding for wasi http
